+++ apps/web-antd/src/views/system/post/data.ts
@@ -82,7 +82,6 @@ export const drawerSchema: FormSchemaGetter = () => [
   {
     component: 'TreeSelect',
     componentProps: {
-      class: 'w-full',
       getPopupContainer,
     },
     fieldName: 'deptId',